moving old code to cscript_old folders
