+++ CMakeLists.txt
@@ -11,8 +11,8 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 #set(CMAKE_CXX_EXTENSIONS OFF)
 
 
-add_subdirectory(cscript)
-add_subdirectory(cscript.tests)
+add_subdirectory(cscript_old)
+add_subdirectory(cscript_old.tests)
 add_subdirectory(vm)
 add_subdirectory(vm.tests)
