+++ cscript_old.tests/optimizetests.h
