+++ cscript_old/compiler.cpp
