+++ cscript_old/cscript.cpp
