+++ cscript_old/defines.h
