+++ cscript_old/namespace.h
