+++ cscript_old/optimize.cpp
