+++ cscript_old/optimize.h
