+++ cscript_old/peephole.cpp
