+++ cscript_old/utility.cpp
