+++ cscript_old/utility.h
